// needrelay-frontend/src/api/generated/models/needStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type NeedStatus = typeof NeedStatus[keyof typeof NeedStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const NeedStatus = {
  OPEN: 'OPEN',
  PARTIALLY_COVERED: 'PARTIALLY_COVERED',
  COVERED: 'COVERED',
  CLOSED: 'CLOSED',
} as const;