// needrelay-frontend/src/api/generated/models/offerStatus.ts
/**
 * Generated by orval v7.21.0 🍺
 * Do not edit manually.
 * NeedRelay API
 * Crisis and emergency resource coordination API
 * OpenAPI spec version: 0.1.0
 */

export type OfferStatus = typeof OfferStatus[keyof typeof OfferStatus];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const OfferStatus = {
  PENDING: 'PENDING',
  COMING: 'COMING',
  RECEIVED: 'RECEIVED',
  CANCELLED: 'CANCELLED',
} as const;